Implemented savings Dao, transfer button, error handling

+++ network/src/main/java/com/bogtech/network/feed/model/subtypes/Amount.kt
@@ -2,5 +2,5 @@ package com.bogtech.network.feed.model.subtypes
 
 data class Amount (
     val currency: String,
-    val minorUnits: Double,
+    val minorUnits: Long,
 )
